view, likes 타입 Long 으로 변경

+++ src/main/java/com/ani/taku_backend/post/model/entity/Post.java
@@ -38,8 +38,8 @@ public class Post {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
-    private int views;
-    private int likes;
+    private Long views;
+    private Long likes;
 
     // 조회수 증가
     public void addViews() {
